requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
djangorestframework-simplejwt>=5.0,<6.0
PyJWT>=2.0,<3.0